save experiment report to XML

+++ scalability_explorer/src/main/java/eu/choreos/vv/data/ExperimentReport.java
@@ -3,7 +3,7 @@ package eu.choreos.vv.data;
 import java.util.HashMap;
 import java.util.List;
 
-public class ExperimentReport extends HashMap<Double, ReportData> {
+public class ExperimentReport extends HashMap<Integer, ReportData> {
 	private static final long serialVersionUID = -2993273004741518768L;
 	private String name;
 	List<String> parameterLabels;
